move KPIs inside netlist too
